fix: replace Float with BigDecimal for all currency values
Float's ~7 significant digits cause rounding errors in exchange rate
calculations. BigDecimal with MathContext.DECIMAL128 (34 digits) gives
exact arithmetic throughout the parse → store → calculate → display
pipeline.

- Rate.value: Float → BigDecimal (all 10 adapters updated)
- SharedPreferences cache: putFloat → putString(toPlainString()); old
  Float entries are skipped on read, triggering a fresh API fetch
- Fee: stored as Float for backward compat, surfaced as BigDecimal
- All arithmetic in MainViewModel uses BigDecimal divide/multiply
- TimelineViewModel avg/min/max/diff use BigDecimal
- TextUtils/MathUtils Float extensions replaced with BigDecimal
- ChartAdapter converts BigDecimal → Float only at the SparkAdapter boundary

+++ app/src/main/kotlin/de/salomax/currencies/model/adapter/InforEuroRatesAdapter.kt
@@ -9,6 +9,7 @@ import de.salomax.currencies.model.Currency
 import de.salomax.currencies.model.ExchangeRates
 import de.salomax.currencies.model.Rate
 import java.io.IOException
+import java.math.BigDecimal
 import java.time.LocalDate
 
 @Suppress("unused", "UNUSED_PARAMETER")
@@ -41,11 +42,11 @@ internal class InforEuroRatesAdapter(private val date: LocalDate) {
     private fun parseEntry(reader: JsonReader): Rate? {
         reader.beginObject()
         var name: Currency? = null
-        var value: Float? = null
+        var value: BigDecimal? = null
         while (reader.hasNext()) {
             when (reader.nextName()) {
                 "isoA3Code" -> name = Currency.fromString(reader.nextString())
-                "value" -> value = reader.nextDouble().toFloat()
+                "value" -> value = reader.nextDouble().toBigDecimal()
                 else -> reader.skipValue()
             }
         }
